Rename core.py to sync.py
A more appropriate/specific name for this particular project.
